requirements.txt
Django==4.0.0
djangorestframework==3.13.1
psycopg2-binary==2.8.5
django-seed==0.3.1
